Debug

+++ src/wr_logic_ai/src/obstacle_avoidance.py
@@ -11,6 +11,7 @@ from sensor_msgs.msg import LaserScan
 from wr_logic_ai.msg import NavigationMsg
 from wr_drive_msgs.msg import DriveTrainCmd
 from geometry_msgs.msg import PoseStamped
+from std_msgs.msg import Float32
 
 import signal
 import sys
@@ -35,7 +36,7 @@ rospy.init_node('nav_autonomous', anonymous=False)
 # Publisher
 drive_pub = rospy.Publisher(
     '/control/drive_system/cmd', DriveTrainCmd, queue_size=1)
-distance_pub = rospy.Publisher('/distance_to_target', float, queue_size=1)
+distance_pub = rospy.Publisher('/distance_to_target', Float32, queue_size=1)
 # TESING
 heading_pub = rospy.Publisher('/debug_heading', PoseStamped, queue_size=1)
 heading_msg = PoseStamped()
@@ -75,7 +76,7 @@ def update_heading_and_target(data) -> None:
     imu = AngleCalculations(data.cur_lat, data.cur_long,
                             data.tar_lat, data.tar_long)
     target_angle = imu.get_angle() % 360
-    distance_pub.publish(imu.get_distance())
+    distance_pub.publish(Float32(imu.get_distance()))
     # TESTING
     print("Current heading: " + str(HEADING))
     print('Target angle: ' + str(target_angle))
